refactor(validations): 공통 유효성 검증 함수 분리
