+++ src/validations/LottoPurchasePriceValidations.js
@@ -1,16 +1,5 @@
-import { COMMON_ERRORS, VALIDATION_ERRORS, LOTTO_PRICE_PER_TICKET, MATCH_PRICE, MATCH_CODE } from '../constants/constants.js';
-
-const validateNumber = (price) => {
-  if (isNaN(price)) {
-    throw new Error(COMMON_ERRORS.NUMBER);
-  }
-};
-
-const validateInteger = (price) => {
-  if (!Number.isInteger(price)) {
-    throw new Error(COMMON_ERRORS.INTEGER);
-  }
-};
+import { VALIDATION_ERRORS, LOTTO_PRICE_PER_TICKET, MATCH_PRICE, MATCH_CODE } from '../constants/constants.js';
+import { validateNumber, validateInteger } from './CommonValidations.js';
 
 const validateLottoAmountRange = (price) => {
   if (price < LOTTO_PRICE_PER_TICKET || price > MATCH_PRICE[MATCH_CODE.SIX]) {
@@ -24,11 +13,11 @@ const validateIsThousandUnit = (price) => {
   }
 };
 
-const validateLottoPurchasePrice = (price) => {
+const LottoPurchasePriceValidations = (price) => {
   validateNumber(price);
   validateInteger(price);
   validateLottoAmountRange(price);
   validateIsThousandUnit(price);
 };
 
-export default validateLottoPurchasePrice;
+export default LottoPurchasePriceValidations;
